Move view card
