+++ app/src/components/market/profile/buy.tsx
@@ -8,7 +8,7 @@ import { Button, BigNumberInput, OutcomeTable } from '../../common'
 import { ERC20Service, MarketMakerService } from '../../../services'
 import { SubsectionTitle } from '../../common/subsection_title'
 import { Table, TD, TR } from '../../common/table'
-import { ViewCard } from '../view_card'
+import { ViewCard } from '../../common/view_card'
 import { computePriceAfterTrade } from '../../../util/tools'
 import { getContractAddress } from '../../../util/addresses'
 import { getLogger } from '../../../util/logger'
