+++ app/src/components/market/profile/sell.tsx
@@ -12,7 +12,7 @@ import { FormRow } from '../../common/form_row'
 import { SubsectionTitle } from '../../common/subsection_title'
 import { Table, TD, TR } from '../../common/table'
 import { TextfieldCustomPlaceholder } from '../../common/textfield_custom_placeholder'
-import { ViewCard } from '../view_card'
+import { ViewCard } from '../../common/view_card'
 import { MarketMakerService } from '../../../services'
 import { useConnectedWeb3Context } from '../../../hooks/connectedWeb3'
 import { getLogger } from '../../../util/logger'
