+++ app/src/components/market/profile/view.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import styled, { withTheme } from 'styled-components'
 import { ethers } from 'ethers'
-import { ViewCard } from '../view_card'
+
+import { ViewCard } from '../../common/view_card'
 import { Status, BalanceItem } from '../../../util/types'
 import { Button } from '../../common'
 import { FullLoading } from '../../common/full_loading'
